Finished JWT authorization

+++ coursework/mern/backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 import expressAsyncHandler from 'express-async-handler'
 import Goal from '../models/goalModel.js'
+import User from '../models/userModel.js'
 
 const getGoals = expressAsyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -14,10 +15,11 @@ const createGoal = expressAsyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
-    res.status(201).json({ message: 'Goal created' })
+    res.status(201).json(goal)
 })
 
 const updateGoal = expressAsyncHandler(async (req, res) => {
@@ -26,6 +28,17 @@ const updateGoal = expressAsyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { returnDocument: 'after' })
@@ -38,6 +51,17 @@ const deleteGoal = expressAsyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
